fix(ci): satisfy newer clippy lints
Remove obsolete cache prune cast and use checked division for cache page alignment so the GitHub runner's current clippy gate stays green.

+++ src/cache.rs
@@ -454,7 +454,7 @@ impl SqlitePairwiseCache {
                         "DELETE FROM pairwise_cache WHERE updated_at < ?1",
                         params![cutoff],
                     )?;
-                    deleted = deleted.saturating_add(removed as usize);
+                    deleted = deleted.saturating_add(removed);
                 }
 
                 if let Some(max_rows) = max_rows {
